tests/wpt/web-platform-tests/css/cssom-view/matchMediaAddListener-handleEvent.html

<!doctype html>
<meta charset="utf-8">
<title>CSS Test: CSSOM View matchMedia handleEvent via addListener</title>
<link rel="help" href="https://dom.spec.whatwg.org/#callbackdef-eventlistener">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="resources/matchMedia.js"></script>
<div id="log"></div>
<script>
setup({ allow_uncaught_exception: true });

promise_test(async t => {
    const iframe = await createIframe(t);
    const mql = iframe.contentWindow.matchMedia(MEDIA_QUERY);

    let _this;
    let _event;
    const eventListener = {
        handleEvent(event) {
            _this = this;
            _event = event;
        },
    };

    mql.addListener(eventListener);
    triggerMQLEvent(iframe);
    await waitForChangesReported();

    assert_equals(_this, eventListener);
    assert_equals(_event.matches, mql.matches);
    assert_equals(_event.media, MEDIA_QUERY);
}, "calls handleEvent method of event listener");

promise_test(async t => {
    const iframe = await createIframe(t);
    const mql = iframe.contentWindow.matchMedia(MEDIA_QUERY);

    let uncaughtError;
    const errorHandler = event => {
        uncaughtError = event.error;
    };
    window.addEventListener("error", errorHandler);
    t.add_cleanup(() => {
        window.removeEventListener("error", errorHandler);
    });

    const thrownError = { name: "test" };
    mql.addListener({
        get handleEvent() {
            throw thrownError;
        },
    });

    triggerMQLEvent(iframe);
    await waitForChangesReported();
    assert_equals(uncaughtError, thrownError);
}, "rethrows errors when getting handleEvent");

promise_test(async t => {
    const iframe = await createIframe(t);
    const mql = iframe.contentWindow.matchMedia(MEDIA_QUERY);

    let calls = 0;
    mql.addListener({
        get handleEvent() {
            calls++;
            return function() {};
        },
    });
    assert_equals(calls, 0);

    triggerMQLEvent(iframe);
    await waitForChangesReported();
    assert_equals(calls, 1);

    triggerMQLEvent(iframe);
    await waitForChangesReported();
    assert_equals(calls, 2);
}, "looks up handleEvent method on every event dispatch");

promise_test(async t => {
    const iframe = await createIframe(t);
    const mql = iframe.contentWindow.matchMedia(MEDIA_QUERY);

    let calls = 0;
    const eventListener = function() { calls++; };
    Object.defineProperty(eventListener, "handleEvent", {
        get: t.unreached_func("handleEvent method should not be looked up on functions"),
    });
    mql.addListener(eventListener);

    triggerMQLEvent(iframe);
    await waitForChangesReported();
    assert_equals(calls, 1);
}, "doesn't look up handleEvent method on callable event listeners");

promise_test(async t => {
    const iframe = await createIframe(t);
    const mql = iframe.contentWindow.matchMedia(MEDIA_QUERY);

    let uncaughtError;
    const errorHandler = event => {
        uncaughtError = event.error;
    };
    window.addEventListener("error", errorHandler);
    t.add_cleanup(() => {
        window.removeEventListener("error", errorHandler);
    });

    mql.addListener({ handleEvent: null });
    triggerMQLEvent(iframe);
    await waitForChangesReported();

    assert_equals(uncaughtError.name, "TypeError");
}, "throws if handleEvent is falsy and not callable");

promise_test(async t => {
    const iframe = await createIframe(t);
    const mql = iframe.contentWindow.matchMedia(MEDIA_QUERY);

    let uncaughtError;
    const errorHandler = event => {
        uncaughtError = event.error;
    };
    window.addEventListener("error", errorHandler);
    t.add_cleanup(() => {
        window.removeEventListener("error", errorHandler);
    });

    mql.addListener({ handleEvent: "str" });
    triggerMQLEvent(iframe);
    await waitForChangesReported();

    assert_equals(uncaughtError.name, "TypeError");
}, "throws if handleEvent is thruthy and not callable");
</script>
